1. gateway filter config 2. global filter to process response 3. create port order for vortex
