+++ vortex-app/vortex-app-api/src/main/resources/application.yaml
@@ -62,6 +62,7 @@ spring:
                 keyName: Authorization
                 adminKey: Bearer ${app.gateway.consoleconnect.admin-api-key}
                 userKey: Bearer ${app.gateway.consoleconnect.user-api-key}
+            - ResponseAdapter
         - id: downstream-api
           uri: ${app.gateway.downstream.base-url}
           predicates:
@@ -115,7 +116,7 @@ logging:
 app:
   route-adapters:
     - method: PUT
-      route-path: /consoleconnect/api/company/consolecore-poping-company/ports/orders
+      route-path: /consoleconnect/api/company/{app.gateway.consoleconnect.company-name}/ports/orders
       adapter-class-name: com.consoleconnect.vortex.gateway.adapter.cc.PortOrderCreateAdapter
   iam:
     jwt:
@@ -178,6 +179,7 @@ app:
       base-url: ${GATEWAY_DOWN_STREAM_BASE_URL:https://api.test.consoleconnect.com}
       admin-api-key: ${GATEWAY_DOWN_STREAM_ADMIN_API_KEY:api-key-has-admin-role}
       user-api-key: ${GATEWAY_DOWN_STREAM_USER_API_KEY:api-key-has-readonly-role}
+      company-name: consolecore-poping-company
     downstream:
       base-url: ${GATEWAY_DOWN_STREAM_BASE_URL:http://localhost:3000}
       admin-api-key: ${GATEWAY_DOWN_STREAM_ADMIN_API_KEY:api-key-has-admin-role}
